+++ src/util/wrap.js
@@ -1,39 +1,45 @@
-import * as d3 from "d3";
+import { select } from "d3";
 
 const length = el => el.node().getComputedTextLength();
 
 // This function is adapted from a Mike Bostock implementation of text wrapping
 // https://bl.ocks.org/mbostock/7555321
-export const wrap = (selection, w) => {
-  let maxTextL = 0, maxSpanL = 0, l;
+const wrap = (selection, w) => {
+  let maxTextL = 0;
+  let maxSpanL = 0;
+  let l;
 
   const lencomp = (text, context) => {
-    if ((l = length(text)) > context) {
-      context = l;
-    }
-    return context;
+    l = length(text);
+    return Math.max(l, context);
   }
 
+  // eslint-disable-next-line func-names
   selection.each(function () {
-    let text = d3.select(this);
+    const text = select(this);
 
     // get the maximum possible length of a label
     maxTextL = lencomp(text, maxTextL);
 
-    let words = text.text().split(/\s+/).reverse();
+    const words = text.text().split(/\s+/).reverse();
     if (words.length === 1) return;
     let word;
-    let line = Array();
-    let x = text.attr("x"), y = text.attr("y");
+    let line = [];
+    const x = text.attr("x");
+    const y = text.attr("y");
     let tspan = text
       .text(null)
       .append("tspan")
       .attr("x", x)
       .attr("y", y);
+
+    // eslint-disable-next-line no-cond-assign
     while (word = words.pop()) {
       line.push(word);
       tspan.text(line.join(" "));
-      if ((l = length(tspan)) > w) {
+
+      l = length(tspan);
+      if (l > w) {
         text.attr("y", text.attr("y") - 5);
         line.pop();
         tspan.text(line.join(" "));
@@ -56,3 +62,5 @@ export const wrap = (selection, w) => {
 
   return (maxSpanL && maxSpanL < maxTextL) ? maxSpanL : maxTextL;
 }
+
+export default wrap;
